Finished questionnnaire form answer storage.
Made sure that questionnaire forms are now storing the answers. Tuned
the database a bit by adding the submitted form identifier (might not be
globally unique, but should be unique enough for usage within the same
session), also changed the sotrage size for session and form identifiers
(just in case there are ever values that would not fit into the filed).
Also added a fw fixes and tuned here and there.