References #3 and #5.
Moved most of the logic into a service, eliminating the need of global
functions. Also dealt with some of the Depedency Injection cases, a few
more are still left. NB! Make sure to uninstall and reistall before
making an update ad there was a typo in allowed values function name for
session template field.