References #10, dashboard is now using a real data.
Number of questies is sub optimal and might require either fetching
larger chunks of data with one query instead of goinf that on
per-questionnaire or per-question basis. Moving some of the processing
(fetching data for visuals into a standalone endpoint and then fetching
that with ajax call might also be required in the future).