References #10, refactoring and optimizations.
Made sure that counts for unique questionnaire submissions are retrieved
with a single query. Left the original method as it was, though it is no
longer being used. Added a generally usable method to re-key results for
counts. The code shold be moved to a service as the Controller class is
getting crowded with additional helper methods and not just the ones
that are serving routes.