References #12, all functionalities are implemented.
Includes implementation for storing user identifier within provided
answers, downloading those as part of the CSV file, providing a filed
for Session Entity to only allow authenticated users and redirecting
users to login page if anonymous and an authenticated one is required.
The implementation for redirection should be centralised and most
probably using the event system to determine it early in the call stack.
There is also a need to redirect the user to the page that triggered
initial redirect after a successful login. Also implemented real
protection from form being submitted instead of just disabling the
submit button.