References #12, implemented redirect back logic.
Implemented logic that would redirect user to a Session Entity page
after login. The implementation uses a short lived cookie that is
httpOnly and has the identifier of the entity. Current lifespan is set
to one hour. This would also allow it to work if there is an account
creation procedure involved. A small side effect is that the redirect
logic would be used for as long as the cookie is valid.