References #19, initial solution.
Added a field to the SessinEntity that has constraints, validators and
is preset with unique code value using callback. The filed would also
get the index set, though that would not be a unique one as that raises
issues with adding filed with migrations in case any entities already
exist. The numeric code is shown on both listing and full view page.
Current implementation allows user to change the code during the
creation phase with unique pregenerated value being preset. NB! Current
implementation does not have the unique value constraint on the database
side and could thus have duplicate values in certain edge cases.
Although that is unlikely.