References #18, implemented the name/nickname requirement.
Initial implementation for name/nickname requirement in case of
anonymous user, provided that the setting requires that. Added
corresponding field to Session Entity and also extended the answers
table with a new column to hold the value. Name would be shown to the
user, if set and session entity requires one. It would be omitted
otherwise.