Reslves #19, added custom postSave method.
Implemented postSave method and added custom code to make sure that the
code is unique. This is present to deal with extremely rare edge cases
of entity with the same code being added into the database after the
validation took place. If that is determined to be the case, then a new
code would be generated instead. This is only applicable of initial
postSave and would be ignored on later updates.