Resolves #28, implemented external dashboard functionality.
Current implementation also has a hidden capability of showing both
external content within an iframe and default dashboard implementation
below it. This requires adding ?show-default-dashboard=true to the
dashboard URL. It is not being used in the codebase and could currently
be manually applied, if needed. Session template preview will notify the
viewer that current template is using a custom dashboard implementation.
Changed the getSessionTemplateData implementation to check if template
could be found. This should prevent errors from being triggered in
listings and views (this should not be possible unless the database is
being changed manually, still, it makes sense to handle that situation).