Changed timers management page.
Added a disabled checkbox element to display current timer status.
Extended the add and edit forms to include status checkbox element.
Removes all the logic for timer-session from management page as timers
would be only used within a context of a session. Timer session entity
would either be removed or repurposed, the logic would be rather
different and current solution would be of not much use.