Implemented missing logic.
Added session entity close hook that is used to stop any timers that are
still running. Implemented removal of all timer data along with the
session entity. Current implementation uses the API and calling the
delete() method on entities directly. That could become an issue if
there are too many entities to be deleted and would require a better
solution. Changed timer export to export the data for session entity
instead of a single timer. Other changes are also possible. This should
complete have all the logic required for the timers, except for exposing
a link to the timers page. The code would need to be refactored and
duplicated code to be moved into a service, permissions and checks are
also required in most of the functions and methods.