Quick feedback is almost fully functional.
Multiple changes to the codebase: added manager, introduced a trait that
is being used by the different questionnaire forms, added questionnaire
entity type, implemented hooks that would allow quick feedback to be
added to the Session Entity. Also includes a lot of other changes. The
codebase is still missing docstrings, some interfaces are not updated
with all the corresponding methods. The whole functionality would need
to be tested before being merged into the main branch.