References #47, mostly implemented + a few other fixes
Implemented the new listing page and made sure that it is mostly used
instead of the default one for administrators. Made sure that Session
Entity forms are able to work with AJAX that is being used by the newly
added listing. Changed all the URLs to the /sessin/... structure, the
only one using old ways is the admin listing. Applied fixes to how the
::ajaxSave is working by removing the overrides to initial submit
handlers and changing the ::save handler to detect AJAX context, store
entity save result in FormState and not redirect or show messages.
Previous approach created issues with additional submission handlers
running and not getting the stored entity data tahat is required and
dealt with by the default ones. Also added a few more fixes here and
there.