References #48, fixed an issue with package being applied twice.
Hook user_form is also being called for user registration. Made sure
that that specific functionality is only applied when user account is
being edited. Also changed the logic of order for the quick feedbakc
questions and user packages to use identifier instead of creation time
as application of a user package has potential to introduce items
created at the same time.